refactor(sprint8): mailbox_handlers.rs (1912 LOC) → mailbox/ 4 sous-modules
- mailbox/mod.rs (525) : setup, helpers folder_to_mailboxes, EmailListQuery
- mailbox/read_handlers.rs (353) : api_email_attachment_download, api_emails, api_tags
- mailbox/send_handlers.rs (792) : api_send, api_send_undo, api_send_schedule, send_queue_worker, api_send_status
- mailbox/single_handlers.rs (111) : api_email_by_id, api_email_action
- mailbox/drafts_handlers.rs (158) : api_drafts_list/upsert/delete
- main.rs: mod mailbox_handlers → mod mailbox
