+++ src/bin/email_api_dir/main.rs
@@ -50,14 +50,14 @@ use serde::{Deserialize, Serialize};
 mod admin_auth;
 mod auth_handlers;
 mod monitoring_handlers;
-mod mailbox_handlers;
+mod mailbox;
 mod admin_ops;
 mod external_handlers;
 mod external_probe_handlers;
 
 pub use auth_handlers::*;
 pub use monitoring_handlers::*;
-pub use mailbox_handlers::*;
+pub use mailbox::*;
 pub use admin_ops::*;
 pub use external_handlers::*;
